refactor: LoginScreen переименован в SignInScreen, PropertyScreen — в RoomScreen
